benchmarked Utf8 encoding based on Text and BoundedEncoding
The special-cased routine for UTF-8 encoding of Text values is a factor 9x
faster than the generic routine based on 'encodeTextWithB'. This implies that
JSON escaping based on 'BoundedEncoding' has a lot of potential, as it
currently is equally fast despite the drawback of being a lot slower on Utf8
encoding.